Abstract out the replacements.
