+++ lib/boilerplate/modes.js
@@ -1,5 +1,6 @@
 'use strict';
 
+/* eslint-disable valid-jsdoc*/
 /**
  * Generate boilerplate output files for various modes.
  *
@@ -52,13 +53,6 @@ const INDENT = 2;
  *
  * We thus read in this project's `package.json` and slice and dice the
  * dependencies as appropriate for a given boilerplate output mode.
- *
- * @param {string}  content           string content
- * @param {*}       opts              options object
- * @param {string}  opts.name         package name
- * @param {string}  opts.description  package description
- * @param {string}  opts.dir          output directory
- * @returns {Promise<string>}         output string content
  */
 const transformPkg = async (content, { name, description }) => {
   // TODO(bp): pkg.json
@@ -94,6 +88,16 @@ const transformPkg = async (content, { name, description }) => {
     null,
     INDENT
   );
+};
+
+/**
+ * Transform examples `index.js` to boilerplate version.
+ *
+ * Generate a transform function with a custom map of regex replacements
+ * very similar to the inject-loader for the underlying examples.
+ */
+const transformIdx = async () => async (content, { name, description }) => {
+  return content;
 };
 
 // ----------------------------------------------------------------------------
